requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
gunicorn
importlib-metadata==8.4.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
werkzeug==3.0.4
zipp==3.20.1
